Add files via upload
The goal of this assignment is to create a virtual network of routers that send routing data to one another. This assignment models the way that actual dynamic network routing is performed.
Link state routing algorithms work on real networks by sending a "link state packet" around the network, with routers forwarding the packet as appropriate. We will be simulating this packet with instances of another class, that will contain the following information:

ID of the router that originates the LSP.
A sequence number indicating how many LSPs have been sent by the originating router; i.e., each newly originated LSP has a higher sequence number than all previous LSPs. The first sequence number should be 1; ignore the possibility of wraparound.
Time to live, decremented each time the LSP is forwarded; with an initial value of 10.
Either of the following at your option:
A list that indicates each reachable network (indicated by the network name stored in the router's string).
A list that indicates each directly connected router, the network behind each one, and the cost to get to that router.